Added Redis as storage
I want something that will be hosted and backed up by somebody else,
so that this service can be treated as an ephemeral thing and the data
stored safely elsewhere.

Not super happy to be storing `*redis-connection*` and calling
ensure-redis-connection, but this is not a life/death service, so
it'll suffice for now.