Be a bit stricter about the content of posts
The issue that set me off on looking into this is that it tooted a
blog post that had no Lisp related content in the post.

Looking a bit deeper into the reason why it would share a blog post
that had no content, I noticed that html-body-to-text function would
make an additional HTTP GET to fetch the post. But there was no need
for it, since the content of the body of the post is already
available.

That was the first issue.

The second issue was the fact that when the GET request was made, and
subsequently lisp-content-p checked the body of the response, there
was Lisp in the body. The problem is that this request fetched the
whole blog post with all the HTML & JavaScript, which contained the
word 'Lisp'. So this was returning a false positive.

Removing html-body-to-text was the right call. And then using
lisp-content-p on the feed's entry content, would tell me 'the truth'.